+++ example/regex_substituter.dart
@@ -20,7 +20,6 @@ library dart_codemod.example.regex_substituter;
 import 'dart:io';
 
 import 'package:codemod/codemod.dart';
-import 'package:source_span/source_span.dart';
 
 /// Pattern that matches a dependency version constraint line for the `codemod`
 /// package, with the first capture group being the constraint.
@@ -32,31 +31,20 @@ final RegExp pattern = RegExp(
 /// The version constraint that `codemod` entries should be updated to.
 const String targetConstraint = '^1.0.0';
 
-class RegexSubstituter implements Suggestor {
-  @override
-  bool shouldSkip(String sourceFileContents) => false;
+Stream<Patch> regexSubstituter(FileContext context) async* {
+  for (final match in pattern.allMatches(context.sourceText)) {
+    final line = match.group(0);
+    final constraint = match.group(1);
+    final updated = line.replaceFirst(constraint, targetConstraint) + '\n';
 
-  @override
-  Iterable<Patch> generatePatches(SourceFile sourceFile) sync* {
-    final contents = sourceFile.getText(0);
-    for (final match in pattern.allMatches(contents)) {
-      final line = match.group(0);
-      final constraint = match.group(1);
-      final updated = line.replaceFirst(constraint, targetConstraint) + '\n';
-
-      yield Patch(
-        sourceFile,
-        sourceFile.span(match.start, match.end),
-        updated,
-      );
-    }
+    yield Patch(updated, match.start, match.end);
   }
 }
 
-void main(List<String> args) {
-  exitCode = runInteractiveCodemod(
+void main(List<String> args) async {
+  exitCode = await runInteractiveCodemod(
     ['regex_substituter_fixtures/pubspec.yaml'],
-    RegexSubstituter(),
+    regexSubstituter,
     args: args,
   );
 }