+++ test/aggregate_suggestor_test.dart
@@ -14,34 +14,18 @@
 
 @TestOn('vm')
 import 'package:codemod/codemod.dart';
+import 'package:codemod/test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:source_span/source_span.dart';
 import 'package:test/test.dart';
 
-class BaseSuggestor implements Suggestor {
-  @override
-  bool shouldSkip(_) => false;
-
-  @override
-  Iterable<Patch> generatePatches(_) sync* {}
+@override
+Stream<Patch> fooSuggestor(_) async* {
+  yield FooPatch();
 }
 
-class AlwaysSkips extends BaseSuggestor {
-  @override
-  bool shouldSkip(_) => true;
-
-  @override
-  Iterable<Patch> generatePatches(_) => [ShouldBeSkippedPatch()];
-}
-
-class FooSuggestor extends BaseSuggestor {
-  @override
-  Iterable<Patch> generatePatches(_) => [FooPatch()];
-}
-
-class BarSuggestor extends BaseSuggestor {
-  @override
-  Iterable<Patch> generatePatches(_) => [BarPatch()];
+@override
+Stream<Patch> barSuggestor(_) async* {
+  yield BarPatch();
 }
 
 class MockPatch extends Mock implements Patch {}
@@ -53,49 +37,10 @@ class BarPatch extends MockPatch {}
 class ShouldBeSkippedPatch extends MockPatch {}
 
 void main() {
-  group('AggregateSuggestor', () {
-    test('accepts a list of Suggestors', () {
-      final foo = BaseSuggestor();
-      final bar = BaseSuggestor();
-      final aggregate = AggregateSuggestor([foo, bar]);
-      expect(aggregate.aggregatedSuggestors, [foo, bar]);
-    });
-
-    test('accepts an iterable of Suggestors', () {
-      final foo = BaseSuggestor();
-      final bar = BaseSuggestor();
-      final aggregate = AggregateSuggestor([foo, bar].map((s) => s));
-      expect(aggregate.aggregatedSuggestors, [foo, bar]);
-    });
-
-    test('shouldSkip() always returns false', () {
-      final aggregate = AggregateSuggestor([AlwaysSkips()]);
-      expect(aggregate.shouldSkip(''), isFalse);
-    });
-
-    group('generatePatches()', () {
-      test('skips suggestors that return true from shouldSkip()', () {
-        final aggregate = AggregateSuggestor([
-          AlwaysSkips(),
-          FooSuggestor(),
-        ]);
-        final sourceFile = SourceFile.fromString('test');
-        final patches = aggregate.generatePatches(sourceFile);
-        expect(patches, hasLength(1));
-        expect(patches.single, TypeMatcher<FooPatch>());
-      });
-
-      test('should yield patches from each suggestor', () {
-        final aggregate = AggregateSuggestor([
-          FooSuggestor(),
-          BarSuggestor(),
-        ]);
-        final sourceFile = SourceFile.fromString('test');
-        final patches = aggregate.generatePatches(sourceFile).toList();
-        expect(patches, hasLength(2));
-        expect(patches[0], TypeMatcher<FooPatch>());
-        expect(patches[1], TypeMatcher<BarPatch>());
-      });
-    });
+  test('aggregate should yield patches from each suggestor', () async {
+    final suggestor = aggregate([fooSuggestor, barSuggestor]);
+    final context = await fileContextForTest('test.dart', 'test');
+    expect(
+        suggestor(context), emitsInOrder([isA<FooPatch>(), isA<BarPatch>()]));
   });
 }