+++ test/examples/regex_substituter_test.dart
@@ -14,7 +14,7 @@
 
 @TestOn('vm')
 import 'package:codemod/codemod.dart';
-import 'package:source_span/source_span.dart';
+import 'package:codemod/test.dart';
 import 'package:test/test.dart';
 
 final RegExp pattern = RegExp(
@@ -24,31 +24,20 @@ final RegExp pattern = RegExp(
 
 const String targetConstraint = '^1.0.0';
 
-class RegexSubstituter implements Suggestor {
-  @override
-  bool shouldSkip(String sourceFileContents) => false;
+Stream<Patch> regexSubstituter(FileContext context) async* {
+  for (final match in pattern.allMatches(context.sourceText)) {
+    final line = match.group(0);
+    final constraint = match.group(1);
+    final updated = line.replaceFirst(constraint, targetConstraint) + '\n';
 
-  @override
-  Iterable<Patch> generatePatches(SourceFile sourceFile) sync* {
-    final contents = sourceFile.getText(0);
-    for (final match in pattern.allMatches(contents)) {
-      final line = match.group(0);
-      final constraint = match.group(1);
-      final updated = line.replaceFirst(constraint, targetConstraint) + '\n';
-
-      yield Patch(
-        sourceFile,
-        sourceFile.span(match.start, match.end),
-        updated,
-      );
-    }
+    yield Patch(updated, match.start, match.end);
   }
 }
 
 void main() {
   group('Examples: RegexSubstituter', () {
-    test('updates `codemod` version', () {
-      final sourceFile = SourceFile.fromString('''
+    test('updates `codemod` version', () async {
+      final context = await fileContextForTest('pubspec.yaml', '''
 dependencies:
   codemod: ^0.2.0
 ''');
@@ -56,9 +45,8 @@ dependencies:
 dependencies:
   codemod: ^1.0.0
 ''';
-      final patches = RegexSubstituter().generatePatches(sourceFile);
-      expect(patches, hasLength(1));
-      expect(applyPatches(sourceFile, patches), expectedOutput);
+      expectSuggestorGeneratesPatches(
+          regexSubstituter, context, expectedOutput);
     });
   });
 }