+++ test_fixtures/functional/before/codemod_no_patches.dart
@@ -3,16 +3,8 @@ import 'dart:io';
 import 'package:codemod/codemod.dart';
 import 'package:glob/glob.dart';
 
-void main(List<String> args) {
-  exitCode = runInteractiveCodemod(
-      filePathsFromGlob(Glob('**')), NoopSuggestor(),
+void main(List<String> args) async {
+  exitCode = await runInteractiveCodemod(
+      filePathsFromGlob(Glob('**')), (_) async* {},
       args: args);
-}
-
-class NoopSuggestor implements Suggestor {
-  @override
-  bool shouldSkip(_) => true;
-
-  @override
-  Iterable<Patch> generatePatches(_) => [];
 }